https://online.sbis.ru/opendoc.html?guid=6887792e-c0e3-4aa4-bec8-67d887f415f6

+++ app/test.js
@@ -1,142 +1,100 @@
 const fs = require('fs-extra');
 const path = require('path');
+const pMap = require('p-map');
+const logger = require('./util/logger');
+const ModulesMap = require('./util/modulesMap');
+const xml = require('./util/xml');
+const Base = require('./Base');
 
-class Test {
-   constructor() {
+const BROWSER_SUFFIX = '_browser';
+const NODE_SUFFIX = '_node';
 
+let getReportTemplate = () => {
+   return {
+      testsuite:{
+         $: {
+            name:"Mocha Tests",
+            tests:"1",
+            failures:"1",
+            errors:"1"
+         },
+         testcase: []
+      }
+   };
+};
+
+let getErrorTestCase = (name, details) => {
+   return {
+      $: {
+         classname: `[${name}]: Test runtime error`,
+         name: "Some test has not been run, see details",
+         time: "0"
+      },
+      failure: details
    }
-   /**
-    * Возвращает список модулей содержащих юнит тесты
-    * @return {Array}
-    * @private
-    */
-   _getTestModules(name) {
-      if (this._testModulesMap.has(name)) {
-         let result = [];
-         this._testModulesMap.get(name).forEach((moduleName) => {
-            let cfg = this._modulesMap.get(moduleName);
-            result = result.concat(cfg.depends || []);
-            result.push(moduleName);
-         });
-      }
-      return this._getModulesFromMap(name);
-   }
-   /**
-    * Возвращает список репозиториев для тестирования
-    * @param {string} name - Название репозитория в конфиге
-    * @return {Array}
-    * @private
-    */
-   _getTestList() {
-      if (this._testList) {
-         return this._testList;
-      }
-      let tests = [];
-      if (!this._testRep.includes('all')) {
-         this._testRep.forEach((testRep) => {
-            let modules = this._getParentModules(this._getTestModules(testRep));
-            tests.push(testRep);
-            modules.forEach((name) => {
-               let cfg = this._modulesMap.get(name);
-               if (!tests.includes(cfg.rep)) {
-                  tests.push(cfg.rep);
-               }
-            });
-         });
-      } else {
-         tests = Object.keys(this._repos).filter((name) => {
-            return !!this._repos[name].test;
-         });
-      }
-      return this._testList = tests.filter((name) => {
-         return !this._repos[name].onlyLoad
+};
+
+class Test extends Base{
+   constructor(cfg) {
+      super(cfg);
+      this._testReports = new Map();
+      this._resources = cfg.resources;
+      this._ports = cfg.ports;
+      this._repos = cfg.repos;
+      this._workspace = cfg._workspace || cfg.workDir;
+      this._testErrors = {};
+      this._modulesMap = new ModulesMap({
+         store: cfg.store,
+         repos: cfg.repos,
+         testRep: cfg.testRep
       });
    }
-
-   _getModulesFromMap(repName) {
-      let moduels = [];
-      this._modulesMap.forEach(cfg => {
-         if (cfg.rep == repName) {
-            moduels.push(cfg.name);
-         }
-      });
-      return moduels;
-   }
-
-   _getParentModules(modules) {
-      let result = modules.slice();
-      this._modulesMap.forEach(cfg => {
-         if (!result.includes(cfg.name) && cfg.depends.some(dependName => result.includes(dependName))) {
-            result.push(cfg.name);
-         }
-      });
-      if (modules.length  !== result.length) {
-         return this._getParentModules(result);
-      }
-      return result;
-   }
-
-   _getChildModules(modules, path) {
-      let result = [];
-      path = path || [];
-      modules.forEach(name => {
-         if (this._modulesMap.has(name) && !path.includes(name)) {
-            let cfg = this._modulesMap.get(name);
-            let depends = this._getChildModules(cfg.depends, path.concat([name]));
-            result.push(name);
-            result = result.concat(depends.filter((item) => !result.includes(item)));
-         }
-      });
-      return result;
-   }
-
 
    /**
     * Дописывает в отчеты название репозитория
     */
-   prepareReport() {
-      this.log('Подготовка отчетов');
+   async prepareReport() {
+      logger.log('Подготовка отчетов');
+      let promisArray = [];
       this._testReports.forEach((filePath, name) => {
          if (fs.existsSync(filePath)) {
-            const parser = new xml2js.Parser();
-            let xml_string = fs.readFileSync(filePath, "utf8");
             let errorText = '';
             if (this._testErrors[name]) {
                errorText = this._testErrors[name].join('<br/>');
             }
-            parser.parseString(xml_string, (error, result) => {
-               if (error === null) {
-                  if (result.testsuite && result.testsuite.testcase) {
-                     result.testsuite.testcase.forEach((item) => {
-                        item.$.classname = `[${name}]: ${item.$.classname}`;
-                     });
-                  } else {
-                     result = {
-                        testsuite: {
-                           testcase: []
-                        }
+            let readPromise = xml.readXmlFile(filePath).then((result) => {
+               if (result.testsuite && result.testsuite.testcase) {
+                  result.testsuite.testcase.forEach((item) => {
+                     item.$.classname = `[${name}]: ${item.$.classname}`;
+                  });
+               } else {
+                  result = {
+                     testsuite: {
+                        testcase: []
                      }
                   }
-
-                  if (errorText) {
-                     result.testsuite.testcase.push(getErrorTestCase(name, errorText));
-                  }
-
-                  this._writeXmlFile(filePath, result);
                }
-               else {
-                  this.log(error);
+
+               if (errorText) {
+                  result.testsuite.testcase.push(getErrorTestCase(name, errorText));
                }
+
+               xml.writeXmlFile(filePath, result);
+            }).catch(error => {
+               logger.log(error);
             });
+
+            promisArray.push(readPromise)
          }
       });
+      return Promise.all(promisArray);
    }
 
    /**
     * Проверяет наличие отчетов по юнит тестам, если какого-то отчета нет кидает ошибку
     */
    checkReport() {
-      this.log('Проверка существования отчетов');
+      logger.log('Проверка существования отчетов');
       let error = [];
       this._testReports.forEach((path, name) => {
          if (!fs.existsSync(path)) {
@@ -145,29 +103,29 @@ class Test {
          }
       });
       if (error.length > 0) {
-         this.log(`Сгенерированы отчеты с ошибками: ${error.join(', ')}`);
+         logger.log(`Сгенерированы отчеты с ошибками: ${error.join(', ')}`);
       }
-      this.log('Проверка пройдена успешно');
+      logger.log('Проверка пройдена успешно');
    }
 
    _createReport(path) {
-      this._writeXmlFile(path, getReportTemplate());
+      xml.writeXmlFile(path, getReportTemplate());
    }
 
-
-
    _getTestConfig(name, suffix) {
-      const testConfig = require('./testConfig.base.json');
+      const testConfig = require('../testConfig.base.json');
       let cfg = Object.assign({}, testConfig);
       let fullName = name + (suffix||'');
-      cfg.tests = this._testModulesMap.has(name) ? this._testModulesMap.get(name) : name + '_test';
-      cfg.root = this._resources;
-      cfg.htmlCoverageReport = cfg.htmlCoverageReport.replace('${module}', fullName).replace('${workspace}', this._workspace);
-      cfg.jsonCoverageReport = cfg.jsonCoverageReport.replace('${module}', fullName).replace('${workspace}', this._workspace);
-      cfg.report = cfg.report.replace('${module}', fullName).replace('${workspace}', this._workspace);
+      let workspace = path.relative(process.cwd(), this._workspace);
+      cfg.tests = this._modulesMap.getTestModules(name);
+      cfg.root =  path.relative(process.cwd(), this._resources);
+      cfg.htmlCoverageReport = cfg.htmlCoverageReport.replace('${module}', fullName).replace('${workspace}', workspace);
+      cfg.jsonCoverageReport = cfg.jsonCoverageReport.replace('${module}', fullName).replace('${workspace}', workspace);
+      cfg.report = cfg.report.replace('${module}', fullName).replace('${workspace}', workspace);
       this._testReports.set(fullName, cfg.report);
       return cfg;
    }
+
    /**
     * Создает конфиги для юнит тестов
     * @param {String} name - название репозитория в конфиге
@@ -176,8 +134,8 @@ class Test {
     */
    _makeTestConfig(name) {
       let defaultPort = 10025;
-      let configPorts = this._argvOptions.ports ? this._argvOptions.ports.split(',') : [];
-      return Promise.all(this._getTestList().map((name, i) => {
+      let configPorts = this._ports ? this._ports.split(',') : [];
+      return Promise.all(this._modulesMap.getTestList().map((name, i) => {
          return new Promise(resolve => {
             let cfg = this._getTestConfig(name, NODE_SUFFIX);
             fs.outputFileSync(`./testConfig_${name}.json`, JSON.stringify(cfg, null, 4));
@@ -191,18 +149,18 @@ class Test {
       }));
    }
 
-
    async _startNodeTest(name) {
       try {
-         await this._execute(
-            `node node_modules/saby-units/cli.js --isolated --report --config="./testConfig_${name}.json"`,
-            __dirname,
+         await this._shell.execute(
+            `node node_modules/saby-units/cli.js --isolated --report --config="testConfig_${name}.json"`,
+            process.cwd(),
             `test node ${name}`
          );
       } catch (e) {
          this._testErrors[name+NODE_SUFFIX] = e;
       }
    }
+
    /**
     * запускает тесты в браузере
     * @param {String} name - название репозитория в конфиге
@@ -212,45 +170,62 @@ class Test {
    async _startBrowserTest(name) {
       let cfg = this._repos[name];
       if (cfg.unitInBrowser) {
-         this.log(`Запуск тестов в браузере`, name);
+         logger.log(`Запуск тестов в браузере`, name);
          try {
-            await this._execute(
-               `node node_modules/saby-units/cli.js --browser --report --config="./testConfig_${name}InBrowser.json"`,
-               __dirname,
+            await this._shell.execute(
+               `node node_modules/saby-units/cli.js --browser --report --config="testConfig_${name}InBrowser.json"`,
+               process.cwd(),
                `test browser ${name}`
             );
          } catch (e) {
             this._testErrors[name+BROWSER_SUFFIX] = e;
          }
-         this.log(`тесты в браузере завершены`, name);
+         logger.log(`тесты в браузере завершены`, name);
       }
    }
+
    async _setContents(value) {
-      //if (fs.existsSync(path.join(this._resources, 'contents.json'))) {
-      this.log(`Замена buildMode в contents на ${value} путь "${path.join(this._resources, 'contents.js')}"`, 'replace_contents');
-      let contents = await fs.readJson(path.join(this._resources, 'contents.json'), "utf8");
-      contents.buildMode = value;
-      await fs.outputFile(`${path.join(this._resources, 'contents.js')}`, `contents=${JSON.stringify(contents)};`);
-      await fs.outputFile(`${path.join(this._resources, 'contents.json')}`, JSON.stringify(contents));
-      //}
+      if (fs.existsSync(path.join(this._resources, 'contents.json'))) {
+         logger.log(`Замена buildMode в contents на ${value} путь "${path.join(this._resources, 'contents.js')}"`, 'replace_contents');
+         let contents = await fs.readJson(path.join(this._resources, 'contents.json'), "utf8");
+         contents.buildMode = value;
+         if (value === 'debug') {
+            this._buildNumber = contents.buildnumber;
+            contents.buildnumber = '';
+         } else {
+            contents.buildnumber = this._buildNumber;
+         }
+         await fs.outputFile(`${path.join(this._resources, 'contents.js')}`, `contents=${JSON.stringify(contents)};`);
+         await fs.outputFile(`${path.join(this._resources, 'contents.json')}`, JSON.stringify(contents));
+      }
    }
+
    /**
     * Запускает тестирование
     * @return {Promise<void>}
     */
-   async startTest() {
-      await this._setContents('debug');
-      await this._makeTestConfig();
-      await pMap(this._getTestList(), (name) => {
-         this.log(`Запуск тестов`, name);
-         return Promise.all([
-            this._startNodeTest(name),
-            this._startBrowserTest(name)
-         ]);
-      },{
-         concurrency: 4
-      });
-      await this._setContents('release')
+   async _run() {
+      try {
+         logger.log(`Запуск тестов`);
+         await this._modulesMap.build();
+         await this._setContents('debug');
+         await this._makeTestConfig();
+         await pMap(this._modulesMap.getTestList(), (name) => {
+            logger.log(`Запуск тестов`, name);
+            return Promise.all([
+               this._startNodeTest(name),
+               this._startBrowserTest(name)
+            ]);
+         }, {
+            concurrency: 4
+         });
+         await this._setContents('release');
+         await this.checkReport();
+         await this.prepareReport();
+         logger.log(`Тестирование завершено`);
+      } catch (e) {
+         throw new Error(`Тестирование завершено с ошибкой ${e}`);
+      }
    }
 }
